disc_parameter_service/requirements.txt
Flask>=2.2.0,<3.0.0
pymongo>=4.0.0,<5.0.0
spacy>=3.2.0,<3.6.0
openai>=0.27.0
python-dotenv>=0.21.0
numpy==1.23.5
flask-cors